server/requirements.txt
django>=5.0.0
django-cors-headers>=4.3.0
pymongo>=4.7.0
gunicorn>=21.2.0
certifi>=2024.2.2
whitenoise>=6.6.0
platformio
google-generativeai
paho-mqtt>=1.6.0
python-dotenv>=1.0.0
bakalapi
APScheduler>=3.10.0
requests>=2.31.0
weasyprint>=60.0; sys_platform != 'win32'
reportlab>=4.0.0
plotly>=5.0.0
kaleido>=0.2.0
pydantic>=2.0.0
django-ratelimit>=4.1.0

